requirements.txt
Flask==2.2.3
flask_login==0.6.2
flask_sqlalchemy==3.0.2
flask_wtf==1.1.1
SQLAlchemy==2.0.4
SQLAlchemy-Utils==0.40.0
WTForms==3.0.1
qrcode==7.4.2
reportlab==3.6.12
svglib==1.5.1
markdown2==2.4.8
python-dotenv==1.0.0
